fix: LISTEN= 绑定IP行为不正确
修复IP乱监听，现在应该可以正确绑定惹

但是多了一坨 `#[allow(dead_code)]` 好难看 QAQ
