+++ src/handler.rs
@@ -140,6 +140,105 @@ pub async fn proxy_image(
     })
 }
 
+// ============================================================================
+// HTTP Request Handler - exported for use in listener module
+// ============================================================================
+
+use http::header::{CACHE_CONTROL, CONTENT_DISPOSITION, CONTENT_TYPE, LOCATION, USER_AGENT};
+use http::{Request, Response};
+use http_body_util::{BodyExt, combinators::BoxBody, Empty, Full};
+use url::form_urlencoded;
+
+#[allow(dead_code)]
+#[inline]
+fn empty() -> BoxBody<Bytes, hyper::Error> {
+    Empty::<Bytes>::new()
+        .map_err(|never| match never {})
+        .boxed()
+}
+
+#[allow(dead_code)]
+#[inline]
+fn full<T: Into<Bytes>>(chunk: T) -> BoxBody<Bytes, hyper::Error> {
+    Full::new(chunk.into())
+        .map_err(|never| match never {})
+        .boxed()
+}
+
+#[allow(dead_code)]
+#[inline]
+fn response_raw(
+    (bytes, ct): (Bytes, Option<String>),
+) -> Response<BoxBody<Bytes, hyper::Error>> {
+    let mut response = Response::new(full(bytes));
+    if let Some(ct) = ct {
+        response
+            .headers_mut()
+            .insert(CONTENT_TYPE, ct.parse().unwrap());
+    }
+    response
+}
+
+#[allow(dead_code)]
+pub async fn handle(
+    downloader: &Downloader,
+    req: Request<hyper::body::Incoming>,
+) -> Result<Response<BoxBody<Bytes, hyper::Error>>, hyper::Error> {
+    let uri = req.uri();
+    match uri.query() {
+        None => Ok(Response::new(full("OK"))), // healthcheck
+        Some(query) => Ok(
+            match proxy_image(
+                downloader,
+                uri.path(),
+                form_urlencoded::parse(query.as_bytes())
+                    .into_owned()
+                    .collect(),
+                req.headers().get(USER_AGENT).map(|ua| ua.to_str().unwrap()),
+            )
+            .await
+            {
+                Ok(file) => {
+                    let mut response = response_raw((file.bytes, Some(file.content_type)));
+                    response.headers_mut().insert(
+                        CACHE_CONTROL,
+                        "max-age=31536000, immutable".parse().unwrap(),
+                    );
+
+                    let mut content_disposition =
+                        format!("inline; filename=\"{}\"", file.filename.0);
+                    if let Some(filename_encoded) = file.filename.1 {
+                        content_disposition =
+                            format!("{content_disposition}; filename*={filename_encoded}");
+                    }
+                    response
+                        .headers_mut()
+                        .insert(CONTENT_DISPOSITION, content_disposition.parse().unwrap());
+                    response
+                }
+                Err(err) => match err {
+                    ProxyImageError::StatusCodeOnly(status_code) => {
+                        let mut response = Response::new(empty());
+                        *response.status_mut() = status_code;
+                        response
+                    }
+                    ProxyImageError::Redirectable(url) => {
+                        let mut response = Response::new(empty());
+                        *response.status_mut() = StatusCode::FOUND;
+                        response
+                            .headers_mut()
+                            .insert(LOCATION, url.parse().unwrap());
+                        response
+                    }
+                    ProxyImageError::BytesOnly(file) => {
+                        response_raw((file.bytes, file.content_type))
+                    }
+                },
+            },
+        ),
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
